refactor: move history ui modules
