+++ src/historyView/index.html
@@ -19,10 +19,10 @@
       {
         "imports": {
           "./modules/domHandlers.js": "${domHandlersUri}",
-          "./modules/historyEvents.js": "${historyEventsUri}",
-          "./modules/historyRenderer.js": "${historyRendererUri}",
+          "./modules/uiManagers/historyEvents.js": "${historyEventsUri}",
+          "./modules/uiManagers/historyRenderer.js": "${historyRendererUri}",
           "./modules/historyViewState.js": "${historyViewStateUri}",
-          "./modules/searchManager.js": "${searchManagerUri}",
+          "./modules/uiManagers/searchManager.js": "${searchManagerUri}",
           "./modules/messageHandlers.js": "${messageHandlersUri}",
           "@common/webviewState.js": "${webviewStateUri}",
           "@common/webview/commands.js": "${commandsUri}",
